Feat: 중복 저장 처리
Related to #1

+++ src/hooks/useRepositoryStorage.js
@@ -7,6 +7,7 @@ const extractRepoData = ({ id, full_name, description, open_issues_count }) => {
   return { id, full_name, description, open_issues_count };
 };
 
+const isExist = (array, id) => array.some((item) => item.id === id);
 const getStoredValue = async (key) => {
   try {
     const item = await AsyncStorage.getItem(key);
@@ -29,6 +30,10 @@ function useRepositoryStorage() {
     }
     try {
       const newRepoData = extractRepoData(repo);
+      if (isExist(data, newRepoData.id)) {
+        notifyMessage('이미 등록된 저장소입니다.');
+        return;
+      }
       const newStoredRepos = data ? [...data, newRepoData] : [newRepoData];
       await AsyncStorage.setItem(
         ASYNC_STORAGE_KEY,
@@ -39,10 +44,30 @@ function useRepositoryStorage() {
     }
   };
 
+  const filteringArray = (id) => {
+    return data.filter((item) => item.id !== id);
+  };
+
+  const removeRepo = async (target) => {
+    try {
+      const newRepoData = filteringArray(target);
+      await AsyncStorage.setItem(
+        ASYNC_STORAGE_KEY,
+        JSON.stringify(newRepoData),
+      );
+    } catch (e) {
+      console.log(e);
+    }
+  };
+
   return {
     repos: data,
     addRepo: (repos) => {
       addRepo(repos);
+      return mutate();
+    },
+    removeRepo: (repo) => {
+      removeRepo(repo);
       return mutate();
     },
   };
